d-day 경과시 요구사항 관련 기능 막음

+++ src/main/java/com/credential/cubrism/server/studygroup/service/StudyGroupService.java
@@ -20,6 +20,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -207,7 +208,7 @@ public class StudyGroupService {
             throw new CustomException(ErrorCode.STUDY_GROUP_ALREADY_JOINED);
         }
 
-        if(pendingMembersRepository.findByUserAndStudyGroup(currentUser, studyGroup).isPresent()) {
+        if (pendingMembersRepository.findByUserAndStudyGroup(currentUser, studyGroup).isPresent()) {
             throw new CustomException(ErrorCode.STUDY_GROUP_ALREADY_REQUESTED);
         }
 
@@ -396,9 +397,17 @@ public class StudyGroupService {
     // 스터디 그룹 목표 추가
     @Transactional
     public ResponseEntity<MessageDto> addStudyGroupGoal(StudyGroupAddGoalDto dto) {
+        Users currentUser = securityUtil.getCurrentUser();
         Long groupId = dto.getStudyGroupId();
         StudyGroup studyGroup = studyGroupRepository.findById(groupId)
                 .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_NOT_FOUND));
+
+        groupMembersRepository.findByUserAndStudyGroupAndAdmin(currentUser, studyGroup, true)
+                .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_NOT_ADMIN));
+
+        if (dDayPassed(studyGroup)) {
+            throw new CustomException(ErrorCode.STUDY_GROUP_DDAY_PASSED);
+        }
 
         StudyGroupGoal goal = new StudyGroupGoal();
         goal.setGoalName(dto.getGoalName());
@@ -426,6 +435,13 @@ public class StudyGroupService {
     public ResponseEntity<MessageDto> deleteStudyGroupGoal(Long goalId) {
         StudyGroupGoal goal = studyGroupGoalRepository.findById(goalId)
                 .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_GOAL_NOT_FOUND));
+        Users currentUser = securityUtil.getCurrentUser();
+        StudyGroup studyGroup = goal.getStudyGroup();
+        groupMembersRepository.findByUserAndStudyGroupAndAdmin(currentUser, studyGroup, true)
+                .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_NOT_ADMIN));
+        if (dDayPassed(studyGroup)) {
+            throw new CustomException(ErrorCode.STUDY_GROUP_DDAY_PASSED);
+        }
 
         List<UserGoal> userGoals = userGoalRepository.findByStudyGroupGoal(goal);
         userGoalRepository.deleteAll(userGoals);
@@ -475,10 +491,14 @@ public class StudyGroupService {
         StudyGroupGoal goal = studyGroupGoalRepository.findById(dto.getGoalId())
                 .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_GOAL_NOT_FOUND));
 
-        UserGoal userGoal = userGoalRepository.findByUserAndStudyGroupGoal(currentUser, goal).orElseThrow(()->new CustomException(ErrorCode.USER_GOAL_NOT_FOUND));
+        UserGoal userGoal = userGoalRepository.findByUserAndStudyGroupGoal(currentUser, goal).orElseThrow(() -> new CustomException(ErrorCode.USER_GOAL_NOT_FOUND));
 
         StudyGroup studyGroup = studyGroupRepository.findById(dto.getGroupId())
                 .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_NOT_FOUND));
+
+        if(dDayPassed(studyGroup)){
+            throw new CustomException(ErrorCode.STUDY_GROUP_DDAY_PASSED);
+        }
 
         if (!userGoal.getUser().getUuid().equals(currentUser.getUuid())) {
             throw new CustomException(ErrorCode.USER_GOAL_NOT_MATCH);
@@ -548,6 +568,14 @@ public class StudyGroupService {
                 .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_NOT_ADMIN));
 
         UserGoal userGoal = studyGroupGoalSubmit.getUserGoal();
+
+        if(dDayPassed(studyGroup)){
+            userGoal.setStudyGroupGoalSubmit(null);
+            studyGroupGoalSubmitRepository.delete(studyGroupGoalSubmit);
+            return ResponseEntity.status(HttpStatus.OK).body(new MessageDto("D-Day가 지났습니다."));
+        }
+
+
         userGoal.setCompleted(true);
         userGoalRepository.save(userGoal);
 
@@ -569,6 +597,10 @@ public class StudyGroupService {
 
         groupMembersRepository.findByUserAndStudyGroupAndAdmin(currentUser, studyGroup, true)
                 .orElseThrow(() -> new CustomException(ErrorCode.STUDY_GROUP_NOT_ADMIN));
+        if(dDayPassed(studyGroup)){
+            studyGroupGoalSubmitRepository.delete(studyGroupGoalSubmit);
+            return ResponseEntity.status(HttpStatus.OK).body(new MessageDto("D-Day가 지났습니다."));
+        }
 
         studyGroupGoalSubmitRepository.delete(studyGroupGoalSubmit);
 
@@ -651,5 +683,13 @@ public class StudyGroupService {
         StudyGroupEnterDto studyGroupEnterDto = new StudyGroupEnterDto(studyGroup.getGroupName(), members, studyGroupDDayDto);
 
         return ResponseEntity.status(HttpStatus.OK).body(studyGroupEnterDto);
+    }
+
+    //스터디 그룹의 dday가 지났는지 확인
+    public boolean dDayPassed(StudyGroup studyGroup) {
+        LocalDate today = LocalDate.now();
+        LocalDate dDay = studyGroup.getDDay().getDDay();
+
+        return !dDay.isAfter(today);
     }
 }
